put sparsifyFunc into its own file
